fix delete tmp dir

+++ scripts/deploy/kuscia.sh
@@ -523,7 +523,7 @@ function start_kuscia_container() {
   if [[ ${IMPORT_SF_IMAGE} = "none"  ]]; then
     echo -e "${GREEN}skip importing sf image${NC}"
   elif [[ ${IMPORT_SF_IMAGE} = "secretflow"  ]]; then
-    if [[ "$domain_type" != "master" ]]; then
+    if [[ "$domain_type" != "master" ]] && [[ ${runtime} == "runc" ]]; then
       docker run --rm $KUSCIA_IMAGE cat ${CTR_ROOT}/scripts/deploy/register_app_image.sh > ${DOMAIN_WORK_DIR}/register_app_image.sh && chmod u+x ${DOMAIN_WORK_DIR}/register_app_image.sh
       bash ${DOMAIN_WORK_DIR}/register_app_image.sh -c ${domain_ctr} -i ${SECRETFLOW_IMAGE} --import
       rm -rf ${DOMAIN_WORK_DIR}/register_app_image.sh
@@ -596,7 +596,7 @@ function start_data_proxy() {
    local counter=0
    local kusciaapi_endpoint="http://localhost:8082/api/v1/serving"
    # import dataproxy image
-   if [[ "$deploy_mode" != "master" ]] && [[ ${runtime} == "runc" ]]; then
+   if [[ "$deploy_mode" != "master" ]]; then
       docker run --rm $KUSCIA_IMAGE cat ${CTR_ROOT}/scripts/deploy/register_app_image.sh > ${DOMAIN_WORK_DIR}/register_app_image.sh && chmod u+x ${DOMAIN_WORK_DIR}/register_app_image.sh
       bash ${DOMAIN_WORK_DIR}/register_app_image.sh -c ${domain_ctr} -i ${DATAPROXY_IMAGE} --import
       rm -rf ${DOMAIN_WORK_DIR}/register_app_image.sh
